+++ spec/unit/profile_parser_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-require "#{PROJECT_ROOT}/profile-parser.rb"
+require "#{PROJECT_ROOT}/puppet-profile-parser.rb"
 
 describe PuppetProfiler do
   it { is_expected.to be_a(Module) }